chore: Parametize namespace and application domain
Parametize all things namespace so we can deploy Discovery
to multiple namespaces. Also make sure we prompt for the
constructed APPLICATION_DOMAIN as the default is CRC and
we need to be able to target any OpenShift.

Luckily the OpenShift service local routes are always
svc or svc.cluster.local so we don't need to updated
the application to fetch the ENV provided <service>_HOST
IP addresses.