feat: Add support for DJANGO_SECRET_KEY
This adds support for DJANGO_SECRET_KEY, defaulting to
a template provided one. the DJANGO_SECRET_KEY is saved in
a discovery-secrets secret as "django-secret-key" and is referenced
as such.

With the DJANGO_SECRET_KEY defined in the OCP deployments,
we now define the PATH as /var/data/secret.txt as we'd be
writing to that location which is the writable data PV.